add function file to src folder, call(import) the funcitons from the scripts folder
